+++ src/program/types/Property.js
@@ -3,11 +3,11 @@ import reserved from '../../utils/reserved.js';
 
 export default class Property extends Node {
 	transpile ( code, transforms ) {
-		if ( transforms.conciseMethodProperty && this.parent.type !== 'ObjectPattern' ) {
+		if ( transforms.conciseMethodProperty && !this.computed && this.parent.type !== 'ObjectPattern' ) {
 			if ( this.shorthand ) {
 				code.insertRight( this.start, `${this.key.name}: ` );
 			} else if ( this.method ) {
-				const name = this.findScope( true ).createIdentifier( this.key.name );
+				const name = this.findScope( true ).createIdentifier( this.key.type === 'Identifier' ? this.key.name : this.key.value );
 				if ( this.value.generator ) code.remove( this.start, this.key.start );
 				code.insertLeft( this.key.end, `: function${this.value.generator ? '*' : ''} ${name}` );
 			}