+++ test/test.js
@@ -29,8 +29,12 @@ function repeat ( str, times ) {
 	return result;
 }
 
+const subsetIndex = process.argv.indexOf( '--subset' );
+const subset = ~subsetIndex ? process.argv[ subsetIndex + 1 ].split( ',' ).map( file => `${file}.js` ) : null;
+const subsetFilter = subset ? file => ~subset.indexOf( file ) : () => true;
+
 describe( 'buble', () => {
-	fs.readdirSync( 'test/samples' ).forEach( file => {
+	fs.readdirSync( 'test/samples' ).filter( subsetFilter ).forEach( file => {
 		var samples = require( './samples/' + file );
 
 		describe( path.basename( file ), () => {
@@ -49,6 +53,8 @@ describe( 'buble', () => {
 			});
 		});
 	});
+
+	if ( subset ) return;
 
 	describe( 'cli', () => {
 		fs.readdirSync( 'test/cli' ).forEach( dir => {
@@ -116,6 +122,7 @@ describe( 'buble', () => {
 				assert.deepEqual( err.loc, { line: 1, column: 4 });
 				assert.equal( err.message, 'Unexpected token (1:4)' );
 				assert.equal( err.snippet, `1 : var 42 = nope;\n        ^` );
+				assert.equal( err.toString(), `SyntaxError: Unexpected token (1:4)\n1 : var 42 = nope;\n        ^` );
 			}
 		});
 
@@ -130,6 +137,7 @@ describe( 'buble', () => {
 				assert.equal( err.loc.column, 13 );
 				assert.equal( err.message, 'x is read-only (1:13)' );
 				assert.equal( err.snippet, `1 : const x = 1; x++;\n                 ^^^` );
+				assert.equal( err.toString(), `CompileError: x is read-only (1:13)\n1 : const x = 1; x++;\n                 ^^^` );
 			}
 		});
 	});